Create separate file for TurtleSim widget
